+++ appseed/ca/visual/visual_visual.cpp
@@ -1,20 +1,27 @@
 #include "StdAfx.h"
 
+
 namespace visual
 {
+
 
    visual::visual(::ca::graphics * pdc)
    {
       m_pdc = pdc;
    }
 
+
    visual::~visual(void)
    {
    }
 
+
    void visual::Polygon(const point_array & pointa)
    {
-      m_pdc->Polygon(pointa.get_data(), pointa.get_size());
+      m_pdc->Polygon(pointa.get_data(), (int) pointa.get_size());
    }
 
+
 } // visual
+
+
